Added inactivation of unauthorized users Added message sending for uneditable messages Improved speed by not updating messages alreaady at home in sync process

+++ services/student_databaser/student_databaser.py
@@ -31,21 +31,41 @@ class StudentDatabaser(BaseService):
         self.cursor.execute(f"UPDATE {var.DATABASE_STUDENTS_TABLE} SET {attribute} = ? WHERE user_id = ?", (value, user_id))  # plz, don't do SQL injection on me :(
         self.connection.commit()
 
+    def _edit_group(self, user_id, group, edit):
+        self.cursor.execute(f"SELECT * FROM {var.DATABASE_STUDENTS_TABLE} WHERE user_id = ?", (user_id,))
+        groups = get_chat_dict(self.cursor.fetchone())['groups']
+        if edit == "add":
+            groups.add(group)
+        if edit == "rm":
+            try:
+                groups.remove(group)
+            except ValueError:
+                pass
+        self._edit_database(user_id, 'groups', ",".join(groups))
+        return groups
+
     # Requests
 
-    def _request_get_chats(self, group=None, sort=False):
+    def _request_get_chats(self, group=None, sort=False, only_active=True):
         self.cursor.execute(f"SELECT * FROM {var.DATABASE_STUDENTS_TABLE}")
         if group is None:
             chats = (get_chat_dict(chat) for chat in self.cursor.fetchall())
         else:
             chats = filter(lambda x: group in x['groups'], (get_chat_dict(chat) for chat in self.cursor.fetchall()))
+        if only_active:
+            chats = filter(lambda x: var.INACTIVE_USER_GROUP not in x['groups'], chats)
         if sort:
             chats = sorted(chats, key=lambda x: x["student_infos"]["surname"] if x["student_infos"]["surname"] is not None else "")
         return tuple(chats)
 
     def _request_new_chat(self, user_id, last_message_id):
-        self.cursor.execute(f"INSERT INTO {var.DATABASE_STUDENTS_TABLE} (user_id, last_message_id) VALUES (?, ?)", (user_id, last_message_id))
-        self.connection.commit()
+        try:
+            self.cursor.execute(f"INSERT INTO {var.DATABASE_STUDENTS_TABLE} (user_id, last_message_id) VALUES (?, ?)", (user_id, last_message_id))
+            self.connection.commit()
+        except sql.IntegrityError:
+            log.warning(f"User with user_id {user_id} already in database, re-enabling it")
+            self._edit_group(user_id, var.INACTIVE_USER_GROUP, "rm")
+            self._edit_database(user_id, 'last_message_id', last_message_id)
         self.cursor.execute(f"SELECT * FROM {var.DATABASE_STUDENTS_TABLE} WHERE user_id = ?", (user_id,))
         return get_chat_dict(self.cursor.fetchone())
 
@@ -59,22 +79,17 @@ class StudentDatabaser(BaseService):
         with open(var.FILEPATH_GROUPS_LIST) as file:
             return tuple({"group": i} for i in file.read().split("\n"))
 
-    def _request_edit_group(self, user_id, group, edit):
-        self.cursor.execute(f"SELECT * FROM {var.DATABASE_STUDENTS_TABLE} WHERE user_id = ?", (user_id,))
-        groups = get_chat_dict(self.cursor.fetchone())['groups']
-        if edit == "add":
-            groups.add(group)
-        if edit == "rm":
-            try:
-                groups.remove(group)
-            except ValueError:
-                pass
-        self._edit_database(user_id, 'groups', ",".join(groups))
-        self.send_request(Request("ghislieri_bot", "set_groups", user_id=user_id, groups=groups))
+    def _request_edit_groups(self, user_id, group, edit):
+        return self._edit_group(user_id, group, edit)
 
-    def _request_remove_chat(self, user_id):
-        self.cursor.execute(f"DELETE FROM {var.DATABASE_STUDENTS_TABLE} WHERE user_id = ?", (user_id,))
-        self.connection.commit()
+    def _request_remove_user(self, user_id, hard_remove=False):
+        if hard_remove:
+            self.cursor.execute(f"DELETE FROM {var.DATABASE_STUDENTS_TABLE} WHERE user_id = ?", (user_id,))
+            self.connection.commit()
+            log.warning(f"User with user_id {user_id} removed from database")
+        else:
+            self._edit_group(user_id, var.INACTIVE_USER_GROUP, "add")
+            log.warning(f"User with user_id {user_id} disabled")
 
     # Exit
